Importer (#8)
* wip, need testing

* csv importer

+++ internal/budget/expendature.go
@@ -9,7 +9,7 @@ import (
 type Expenditure struct {
 	Owner          uuid.UUID `db:"owner"`
 	Name           string    `db:"name"`
-	Amount         int       `db:"amount"`
+	Amount         float64   `db:"amount"`
 	Date           time.Time `db:"date"`
 	Method         string    `db:"method"`
 	BudgetCategory string    `db:"budget_category"`
