+++ migrations/000001_create_table_budget.up.sql
@@ -7,14 +7,14 @@ CREATE TABLE IF NOT EXISTS budget (
 CREATE TABLE IF NOT EXISTS income (
     owner UUID,
     source VARCHAR(50),
-    amount INT,
+    amount NUMERIC(20,4),
     PRIMARY KEY (owner, source)
 );
 
 CREATE TABLE IF NOT EXISTS expense (
     budget_id UUID,
     category VARCHAR(20),
-    amount INT,
+    amount NUMERIC(20,4),
     is_fixed BOOLEAN,
     is_slack BOOLEAN,
     PRIMARY KEY (budget_id, category)
